+++ adminapp/src/pages/ProgramForm.jsx
@@ -1,18 +1,13 @@
 import api from "../api";
-import AutocompleteSearch from "../components/AutocompleteSearch";
 import FormLayout from "../components/FormLayout";
 import ImageFileInput from "../components/ImageFileInput";
 import MultiLingualText from "../components/MultiLingualText";
+import OneToManyEditor from "../components/OneToManyEditor";
 import ResponsiveStack from "../components/ResponsiveStack";
 import SafeDateTimePicker from "../components/SafeDateTimePicker";
 import { formatOrNull } from "../modules/dayConfig";
-import mergeAt from "../shared/mergeAt";
-import withoutAt from "../shared/withoutAt";
-import AddIcon from "@mui/icons-material/Add";
-import DeleteIcon from "@mui/icons-material/Delete";
 import RemoveIcon from "@mui/icons-material/Remove";
-import { Box, FormHelperText, FormLabel, Icon, Stack, TextField } from "@mui/material";
-import Button from "@mui/material/Button";
+import { FormHelperText, FormLabel, Stack, TextField } from "@mui/material";
 import React from "react";
 
 export default function ProgramForm({
@@ -120,7 +115,7 @@ export default function ProgramForm({
           fullWidth
           onChange={setFieldFromInput}
         />
-        <ModelItems
+        <OneToManyEditor
           title="Commerce Offering"
           items={resource.commerceOfferings}
           setItems={(o) => setField("commerceOfferings", o)}
@@ -128,69 +123,5 @@ export default function ProgramForm({
         />
       </Stack>
     </FormLayout>
-  );
-}
-
-function ModelItems({ title, items, setItems, apiItemSearch }) {
-  const handleAdd = () => {
-    setItems([...items, { id: 0 }]);
-  };
-  const handleRemove = (index) => {
-    setItems(withoutAt(items, index));
-  };
-  function handleChange(index, fields) {
-    setItems(mergeAt(items, index, fields));
-  }
-  return (
-    <>
-      <FormLabel>{`${title}s`}</FormLabel>
-      {items?.map((o, i) => (
-        <ModelItem
-          key={i + title}
-          {...o}
-          name={o.name || o.description?.en}
-          title={title}
-          index={i}
-          apiItemSearch={apiItemSearch}
-          onChange={(fields) => handleChange(i, fields)}
-          onRemove={() => handleRemove(i)}
-        />
-      ))}
-      <Button onClick={handleAdd}>
-        <AddIcon /> Add {title}
-      </Button>
-    </>
-  );
-}
-
-function ModelItem({ index, title, name, apiItemSearch, onChange, onRemove }) {
-  return (
-    <Box sx={{ p: 2, border: "1px dashed grey" }}>
-      <Stack
-        direction="row"
-        spacing={2}
-        mb={2}
-        sx={{ justifyContent: "space-between", alignItems: "center" }}
-      >
-        <FormLabel>
-          {title} {index + 1}
-        </FormLabel>
-        <Button onClick={onRemove} variant="warning" sx={{ marginLeft: "5px" }}>
-          <Icon color="warning">
-            <DeleteIcon />
-          </Icon>
-          Remove
-        </Button>
-      </Stack>
-      <AutocompleteSearch
-        label={title}
-        value={name || ""}
-        fullWidth
-        required
-        search={apiItemSearch}
-        style={{ flex: 1 }}
-        onValueSelect={(vs) => onChange({ ...vs })}
-      />
-    </Box>
   );
 }
